rename extensions class
